refactor: move login / signup controllers in separate folder
